requirements.txt
numpy==1.24.3
openpyxl==3.0.10
pandas==1.5.3
python-dotenv==1.1.1
selenium==4.16.0
tqdm==4.67.1